Login system in 5 minutes!
This system is part of my code in 5 mintues project.
This system is very insecure and could be greatly improved.
It simply is supposed to give the user a basic understanding of a login system.
In no way do I reccomend anyone use this for a website. Just use parts of it.
This code it under the GNU Public Licence Version 3.
Written in 5 minutes by LaughingQuoll.